+++ Tests/Get.test.ps1
@@ -1,12 +1,83 @@
 
 . ./Zoo.ps1
 
-task AsPSCustomObject {
+task AsPS {
 	Use-LiteDatabase :memory: {
 		$test = Get-LiteCollection Test Int32
 		Add-LiteData $test (New-BsonBag)
-		$r = Get-LiteData $test -As ps
+		$r = Get-LiteData $test -As PS
 		equals $r.GetType().Name PSCustomObject
+		equals $r.double 3.14
+		equals $r.string bar
+		equals $r.object.bar 1
+		equals $r.array[0] 1
+		equals $r.array[1] 2
+		equals $r.binData04 ([guid]"cdccdb76-30a3-4d7c-97fa-5ae1ad28fd64")
+		equals $r.binData00.GetType() ([byte[]])
+		equals $r.objectId ([LiteDB.ObjectId]"5dc4c9808c94b4316c418f95")
+		equals $r.bool $true
+		equals $r.date ([DateTime]"2019-11-11")
+		equals $r.null $null
+		equals $r.int 42
+		equals $r.long 42L
+		equals $r.decimal 123456789.123456789d
+		equals $r.minKey ([LiteDB.BsonValue]::MinValue)
+		equals $r.maxKey ([LiteDB.BsonValue]::MaxValue)
+	}
+}
+
+task AsType {
+	# id ~ _id
+	class T1 {$id; $p1}
+
+	# literal mapping
+	class T2 {$_id; $p1}
+
+	# <ClassName>Id ~ _id
+	class T3 {$T3Id; $p1}
+
+	# no ID
+	class T4 {$p1; $extra}
+
+	Use-LiteDatabase :memory: {
+		$test = Get-LiteCollection test
+
+		Add-LiteData $test @{_id = 33; p1 = 38; extra = 60}
+
+		$r = Get-LiteData $test -As T1
+		equals $r.GetType() ([T1])
+		equals $r.id 33
+		equals $r.p1 38
+
+		$r = Get-LiteData $test -As T2
+		equals $r.GetType() ([T2])
+		equals $r._id 33
+		equals $r.p1 38
+
+		$r = Get-LiteData $test -As T3
+		equals $r.GetType() ([T3])
+		equals $r.T3Id 33
+		equals $r.p1 38
+
+		$r = Get-LiteData $test -As T4
+		equals $r.GetType() ([T4])
+		equals $r.p1 38
+		equals $r.extra 60
+
+		$r = Get-LiteData $test -As ([System.Collections.Generic.Dictionary[string, object]])
+		equals $r.GetType() ([System.Collections.Generic.Dictionary[string, object]])
+		equals $r.Count 3
+		equals $r._id 33
+
+		# https://github.com/mbdavid/LiteDB/pull/1480
+		$r = Get-LiteData $test -As ([object])
+		equals $r.GetType() ([System.Collections.Generic.Dictionary[string, object]])
+		equals $r.Count 0
+
+		# https://github.com/mbdavid/LiteDB/pull/1480
+		$r = Get-LiteData $test -As ([hashtable])
+		equals $r.GetType() ([hashtable])
+		equals $r.Count 0
 	}
 }